requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
google-generativeai
pandas
numpy
python-dotenv
tqdm